增加HttpContentTypeAttribute
